Fixed time limit

+++ Project Quinoa/Assets/Scripts/Gameplay/Health.cs
@@ -58,7 +58,7 @@ public class Health : MonoBehaviour {
                 {
                     RandomMatchmaker nm = GameObject.FindObjectOfType<RandomMatchmaker>();
 				
-					PhotonNetwork.player.customProperties["Deaths"] = PhotonNetwork.player.customProperties["Deaths"] + 1;
+					PhotonNetwork.player.customProperties["Deaths"] = (int)PhotonNetwork.player.customProperties["Deaths"] + 1;
 					PhotonNetwork.player.SetCustomProperties (PhotonNetwork.player.customProperties);
 				
                     nm.standby.SetActive(true);
